IDAM-103 - adding account controller to the nuget to see if it works. (it does).
