+++ TestClients/NICE.Identity.TestClient.NETCore/Views/Shared/_Layout.cshtml
@@ -37,11 +37,18 @@
 	            <ul class="nav navbar-nav navbar-right">
 		            @if (User.Identity.IsAuthenticated)
 		            {
-			            <li><a  asp-controller="Account" asp-action="Logout">Logout</a></li>
+			            <li>@*<a  asp-controller="Account" asp-action="Logout">Logout</a>*@
+			            
+			        <a href="/account/logout">Logout</a>    
+					</li>
 		            }
 		            else
 		            {
-			            <li><a asp-controller="Account" asp-action="Login">Login</a></li>
+						<li>
+							<a href="/account/login">Login</a>
+
+							@*<a asp-controller="Account" asp-action="Login">Login</a>*@
+						</li>
 		            }
 	            </ul>
             </div>
